# docs/swagger.yaml
basePath: /
definitions:
  config.HeuristicWeights:
    properties:
      block_potential:
        description: 70 for blocking 2-in-a-row
        type: integer
      block_when_threat:
        description: Block opponent's path values (context-dependent)
        type: integer
      build_alignment_2:
        description: Formation building (our cards in a row)
        type: integer
      build_alignment_3:
        description: 100 for 3-in-a-row
        type: integer
      keep_near_card:
        description: 60 for placing near own cards
        type: integer
      legal_move:
        description: Base legal move value
        type: integer
      play_smallest_card:
        description: Card management bonuses
        type: integer
      replace_pos_center:
        description: Position bonuses when replacing opponent's card
        type: integer
      replace_pos_side:
        description: 50 for side position
        type: integer
      replace_potential:
        description: 125 when blocking potential threat
        type: integer
      replace_values_potential:
        additionalProperties:
          type: integer
        description: 'Card values for defensive play (low cards preferred: 1→100,
          9→20)'
        type: object
      replace_values_threat:
        additionalProperties:
          type: integer
        description: 'Card values when blocking threat (high cards preferred: 1→20,
          9→100)'
        type: object
      replace_when_threat:
        description: Replace opponent's card values (context-dependent)
        type: integer
      w_threat:
        description: Threat detection (3 opponent cards in a row)
        type: integer
      w_win:
        description: Winning move (4-in-a-row)
        type: integer
    type: object
  http.JoinRoomRequest:
    properties:
      player_name:
        type: string
      room_code:
        type: string
    type: object
  http.PlayRequest:
    properties:
      number_bot:
        type: integer
      number_player:
        type: integer
      player_name:
        description: Changed to array
        items:
          type: string
        type: array
      room_id:
        type: string
      weights:
        $ref: '#/definitions/config.HeuristicWeights'
    type: object
info:
  contact:
    email: backend@yourcompany.com
    name: Backend Team
  description: REST API for heuristic-based chess-like bot (Go + Gin)
  title: Javanese Chess Bot API
  version: "1.0"
paths:
  /api/config/weights/default:
    get:
      description: Returns the default heuristic weights based on research paper (Section
        2.4)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get default heuristic weights
      tags:
      - Config
  /api/config/weights/room:
    get:
      description: Returns the heuristic weights configured for a specific room
      parameters:
      - description: Room Code
        in: query
        name: roomCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get room heuristic weights
      tags:
      - Config
  /api/join:
    post:
      consumes:
      - application/json
      description: Join an existing room with a room code
      parameters:
      - description: Join room info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.JoinRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Join an existing room
      tags:
      - Room
  /api/play:
    post:
      consumes:
      - application/json
      description: Initialize room (create if missing), add bots and apply provided
        heuristic weights in one request
      parameters:
      - description: Room info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.PlayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Add bots to a room or create room and apply config
      tags:
      - Room
swagger: "2.0"